+++ Branch-SDK/src/io/branch/referral/Defines.java
@@ -15,6 +15,7 @@ public class Defines {
         SessionID("session_id"),
         LinkClickID("link_click_id"),
         GoogleSearchInstallReferrer("google_search_install_referrer"),
+        GooglePlayInstallReferrer("install_referrer_extras"),
         FaceBookAppLinkChecked("facebook_app_link_checked"),
         BranchLinkUsed("branch_used"),
         ReferringBranchIdentity("referring_branch_identity"),