+++ ChangeLog.md
@@ -1,4 +1,16 @@
 Branch Android SDK change log
+- v2.10.3 Fix for reading intent from launcher activity only unless forced to restart session. Instant app utility methods fix for Android `O`
+
+- v2.10.2 More reliable install referrer capture. URL encoding for supporting referring links added in play store referrer
+
+- v2.10.1 Deprecating enablePlayStoreReferrer()
+
+- v2.10.0 Added changes to capture raw play store referrer string when Google search install referrer is present
+
+- v2.9.1 Hot fix for a possible app crash on instant app to full app conversion
+
+- v2.9.0 Adding network layer abstraction to build custom network layer. Apps can build their own network implementation to handle Branch network requests. By default Branch will be using URL connection for handling network requests.
+
 - v2.8.0 Adding Instant App support. Supports Instant App content deep linking, Full app conversion with deep linking and associated conversion attributions.
 
 - v2.7.0 Moved `BranchLinkUsed` to intent extra. Fixed ability to simulate is_first_session:true when testing. CD Revamp.